fix for value smoothing and updated debug

+++ convergence_app/src/serialReceiver.h
@@ -35,7 +35,7 @@ public:
     bool touched[ NBOARDS * NPINS ];
     bool lastTouched[ NBOARDS * NPINS ];
     int readings[ NBOARDS * NPINS ];
-    int smoothedReadings[ NBOARDS * NPINS ];
+    float smoothedReadings[ NBOARDS * NPINS ];
     int readingsArrays[ NBOARDS * NPINS ][1000];
     vector< vector<int> > readingsVectors;
     int numReadingsStored, numReadingsSmoothed;
